+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ### [Unreleased]
-*
+* Caching of unchanged resources with ETag and If-None-Match headers
  
 ### [0.5.0] - 2017-10-20
 * Support for HTTP HEAD Method